Changed structure and fixes for the plugin
