+++ php/Config.php
@@ -1,14 +1,14 @@
 <?php
 /**
  * Configuration file for the plugin.
+ *
+ * @package AkshitSethi\Plugins\CustomizeWoo
  */
 
 namespace AkshitSethi\Plugins\CustomizeWoo;
 
 /**
  * Set configuration options.
- *
- * @package AkshitSethi\Plugins\CustomizeWoo
  */
 class Config {
 
@@ -97,8 +97,8 @@ class Config {
 
 			// Authentication
 			'woocommerce_lost_password_confirmation_message' => esc_html( 'A password reset email has been sent to the email address on file for your account, but may take several minutes to show up in your inbox. Please wait at least 10 minutes before attempting another reset.', 'customize-woo' ),
-			'woocommerce_lost_password_message'          => esc_html( 'Lost your password? Please enter your username or email address. You will receive a link to create a new password via email.', 'customize-woo' ),
-			'woocommerce_reset_password_message'         => esc_html( 'Enter a new password below.', 'customize-woo' ),
+			'woocommerce_lost_password_message'          => esc_html__( 'Lost your password? Please enter your username or email address. You will receive a link to create a new password via email.', 'customize-woo' ),
+			'woocommerce_reset_password_message'         => esc_html__( 'Enter a new password below.', 'customize-woo' ),
 
 			// Misc
 			'woocommerce_countries_tax_or_vat'           => esc_html__( 'Tax for USA, VAT for European countries', 'customize-woo' ),
